I added some parameters to the packed_scene from path constructor
Asuming we're treating scene.hpp (well the project in general) as a public framework/library of sorts I thought having access to ResourceLoader parameters through the constructor would be quite convenient.

ResourceLoader has some parameters that are otherwise inaccessible.

Notably the CacheMode, which defaults to 1 = CACHE_MODE_REUSE. This caches previous loads and reuses them, which is pretty neat, but the use of this cache may not always be required, so it's best to give the user the option to change this mode.